requirements.txt
fastapi==0.116.2
uvicorn==0.23.2
tensorflow==2.20.0
numpy==2.3.3
pillow==11.3.0
gdown==5.2.0



